move export and file reader to estimationsSection
