+++ e2e-tests/test-helpers.ts
@@ -1,11 +1,10 @@
 import AxeBuilder from '@axe-core/playwright';
-import type { Page, Locator } from '@playwright/test';
+import type { Page } from '@playwright/test';
 import { expect } from '@playwright/test';
 import { OrganisationSection } from './page-objects/organisation-section';
 import { CloudServicesSection } from './page-objects/cloud-services-section';
 import { EndUsersSection } from './page-objects/end-users-section';
 import { OnPremSection } from './page-objects/on-prem-section';
-import * as fs from 'fs';
 
 export async function assertAllSectionElementsAreVisible(
   organisationSection: OrganisationSection,
@@ -23,26 +22,6 @@ export const expectNoA11yViolations = async (page: Page) => {
   const results = await new AxeBuilder({ page }).analyze();
   expect(results.violations).toEqual([]);
 };
-
-export async function exportJsonContent(page: Page, exportType: 'Export JSON' | 'Export JSON with Inputs') {
-  const [download] = await Promise.all([
-    page.waitForEvent('download'),
-    page.getByRole('button', { name: 'Export ▼' }).click(),
-    page.getByRole('link', { name: exportType, exact: true }).click(),
-  ]);
-
-  const path = await download.path();
-  if (!path) throw new Error('Download failed');
-
-  return path;
-}
-
-export async function readJsonFileContent(path: string) {
-  const fileContent = fs.readFileSync(path, 'utf-8');
-  const json = JSON.parse(fileContent);
-
-  return json;
-}
 
 interface EmissionValuesSchema {
   values: {
@@ -130,12 +109,6 @@ interface EmissionInputsSchema {
     };
   };
 }
-
-// export interface EmissionsData {
-//   values: EmissionsValues;
-//   percentages: EmissionsPercentages;
-//   input?: InputData;
-// }
 
 export function createDefaultValuesJsonExport(overrides: Partial<EmissionValuesSchema> = {}): EmissionValuesSchema {
   const defaultValuesJson = {
